Fastplace 1.1
This is a mod for Minecraft (1.8.9) that allows you to set the block place delay with /setdelay.
To install the mod, simply drag the file Fastplace-1.1-(1.8.9).jar into your mods folder.

Key Authors: Myzo, Nanook
Check him out: https://github.com/NotNanook

Changes from 1.0: Added error messages if the input is not int or missing.

Changes from 1.1: Made it so that it only works when holding a block.